feat: Add index selection functionality to API endpoint
This commit adds the functionality to select different indices (NDVI, NDWI, or default) when making an API call to the Copernicus service. The `indexSelection` function in the `config/index-selection.js` file has been implemented to return the appropriate evalscript based on the selected index. The selected index is obtained from the query parameters in the API request.

The evalscript is then used in the API request body to process the satellite imagery data and generate the corresponding output. The generated output is saved as a TIFF file in the `resources` directory, with the file name based on the selected index. The file is then sent as a response to the client.

This feature enhances the flexibility of the API by allowing users to choose the index they want to apply to the satellite imagery data.